// apps/server/src/generated/prisma/models/SystemNotification.ts
export type SystemNotificationModel = runtime.Types.Result.DefaultSelection<Prisma.$SystemNotificationPayload>
// ...
export type $SystemNotificationPayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "SystemNotification"
  objects: {}
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: string
    userId: string
    userName: string | null
    module: string
    type: string
    title: string
    content: string
    targetType: string | null
    targetId: string | null
    routePath: string | null
    status: string
    shownAt: Date | null
    readAt: Date | null
    metadata: runtime.JsonValue | null
    createdAt: Date
    updatedAt: Date
  }, ExtArgs["result"]["systemNotification"]>
  composites: {}
}